flask_photo_app/photo_app/requirements.txt
requests
flask
Flask-WTF
Pillow
flask-login
python_jose
boto3
#flask-lambda
#flask_serverless